Extract logging to its own subpackage.
